mcp/requirements.txt
mcp[cli]>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
